fix(mission): exp 타입 int로 변경
- exp가 현재 100까지밖에 없기 때문에

+++ backend/choki/src/main/java/com/yeojiphap/choki/domain/mission/dto/MissionAddRequestDto.java
@@ -12,5 +12,5 @@ public class MissionAddRequestDto {
 	public Long childId;
 	public MissionType missionType;
 	public String content;
-	public Long exp;
+	public int exp;
 }
